fix: expose way to deserialize to internal object from external IJsonPluggableLibrary implementations, remove duplicate/non-IJsonPluggableLibrary-based deserialization logic
