+++ src/Api/PubnubApi/EndPoint/PubSub/SubscribeManager2.cs
@@ -114,7 +114,7 @@ namespace PubnubApi.EndPoint
             if (!string.IsNullOrEmpty(responseTuple.Item1) && responseTuple.Item2 == null)
             {
                 PNStatus status = new PNStatus(null, PNOperationType.PNSubscribeOperation, PNStatusCategory.PNConnectedCategory, channels, channelGroups);
-                HandshakeResponse handshakeResponse = JsonConvert.DeserializeObject<HandshakeResponse>(responseTuple.Item1);
+                HandshakeResponse handshakeResponse = jsonLibrary.DeserializeToObject<HandshakeResponse>(responseTuple.Item1);
                 return new Tuple<HandshakeResponse, PNStatus>(handshakeResponse, status);
             }   
 
@@ -170,7 +170,7 @@ namespace PubnubApi.EndPoint
                 if (!string.IsNullOrEmpty(responseTuple.Item1) && responseTuple.Item2 == null)
                 {
                     PNStatus status = new PNStatus(null, PNOperationType.PNSubscribeOperation, PNStatusCategory.PNConnectedCategory, channels, channelGroups);
-                    ReceivingResponse<object> receiveResponse = JsonConvert.DeserializeObject<ReceivingResponse<object>>(responseTuple.Item1);
+                    ReceivingResponse<object> receiveResponse = jsonLibrary.DeserializeToObject<ReceivingResponse<object>>(responseTuple.Item1);
                     return new Tuple<ReceivingResponse<object>, PNStatus>(receiveResponse, status);
                 }
                 else if (responseTuple.Item2 != null)
