+++ src/Api/PubnubApi/JsonDataParse/PNGrantTokenJsonDataParse.cs
@@ -1,21 +1,18 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
+﻿using System.Collections.Generic;
 
 namespace PubnubApi
 {
     internal static class PNGrantTokenJsonDataParse
     {
-        internal static PNAccessManagerTokenResult GetObject(List<object> listObject)
+        internal static PNAccessManagerTokenResult GetObject(IJsonPluggableLibrary jsonPlug, List<object> listObject)
         {
             PNAccessManagerTokenResult result = null;
 
-            Dictionary<string, object> grantDicObj = JsonDataParseInternalUtil.ConvertToDictionaryObject(listObject[0]);
+            Dictionary<string, object> grantDicObj = jsonPlug.ConvertToDictionaryObject(listObject[0]);
             if (grantDicObj != null && grantDicObj.ContainsKey("data"))
             {
                 result = new PNAccessManagerTokenResult();
-                Dictionary<string, object> grantDataDic = JsonDataParseInternalUtil.ConvertToDictionaryObject(grantDicObj["data"]);
+                Dictionary<string, object> grantDataDic = jsonPlug.ConvertToDictionaryObject(grantDicObj["data"]);
                 if (grantDataDic != null && grantDataDic.ContainsKey("token"))
                 {
                     result.Token = grantDataDic["token"].ToString();
