+++ src/Api/PubnubApi/JsonDataParse/PNPublishFileMessageJsonDataParse.cs
@@ -1,7 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 
 namespace PubnubApi
 {
