+++ src/Api/PubnubApi/JsonDataParse/PNRemoveMessageActionJsonDataParse.cs
@@ -1,15 +1,12 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
+﻿using System.Collections.Generic;
 
 namespace PubnubApi
 {
     internal static class PNRemoveMessageActionJsonDataParse
     {
-        internal static PNRemoveMessageActionResult GetObject(List<object> listObject)
+        internal static PNRemoveMessageActionResult GetObject(IJsonPluggableLibrary jsonPlug, List<object> listObject)
         {
-            Dictionary<string, object> removeMsgActionDicObj = JsonDataParseInternalUtil.ConvertToDictionaryObject(listObject[0]);
+            Dictionary<string, object> removeMsgActionDicObj = jsonPlug.ConvertToDictionaryObject(listObject[0]);
             PNRemoveMessageActionResult result = null;
             if (removeMsgActionDicObj != null && removeMsgActionDicObj.ContainsKey("status"))
             {
