+++ src/Api/PubnubApi/JsonDataParse/PNSetChannelMetadataJsonDataParse.cs
@@ -1,21 +1,18 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
+﻿using System.Collections.Generic;
 
 namespace PubnubApi
 {
     internal static class PNSetChannelMetadataJsonDataParse
     {
-        internal static PNSetChannelMetadataResult GetObject(List<object> listObject)
+        internal static PNSetChannelMetadataResult GetObject(IJsonPluggableLibrary jsonPlug, List<object> listObject)
         {
-            Dictionary<string, object> setChannelMetadataDicObj = (listObject != null && listObject.Count >= 2) ? JsonDataParseInternalUtil.ConvertToDictionaryObject(listObject[1]) : null;
+            Dictionary<string, object> setChannelMetadataDicObj = (listObject != null && listObject.Count >= 2) ? jsonPlug.ConvertToDictionaryObject(listObject[1]) : null;
             PNSetChannelMetadataResult result = null;
             if (setChannelMetadataDicObj != null && setChannelMetadataDicObj.ContainsKey("data"))
             {
                 result = new PNSetChannelMetadataResult();
 
-                Dictionary<string, object> getSetChMetadataDataDic = JsonDataParseInternalUtil.ConvertToDictionaryObject(setChannelMetadataDicObj["data"]);
+                Dictionary<string, object> getSetChMetadataDataDic = jsonPlug.ConvertToDictionaryObject(setChannelMetadataDicObj["data"]);
                 if (getSetChMetadataDataDic != null && getSetChMetadataDataDic.Count > 0)
                 {
                     result.Channel = getSetChMetadataDataDic.ContainsKey("id") && getSetChMetadataDataDic["id"] != null ? getSetChMetadataDataDic["id"].ToString() : null;
@@ -24,7 +21,7 @@ namespace PubnubApi
                     result.Updated = getSetChMetadataDataDic.ContainsKey("updated") && getSetChMetadataDataDic["updated"] != null ? getSetChMetadataDataDic["updated"].ToString() : null;
                     if (getSetChMetadataDataDic.ContainsKey("custom"))
                     {
-                        result.Custom = JsonDataParseInternalUtil.ConvertToDictionaryObject(getSetChMetadataDataDic["custom"]);
+                        result.Custom = jsonPlug.ConvertToDictionaryObject(getSetChMetadataDataDic["custom"]);
                     }
                 }
             }
