+++ src/Api/PubnubApi/PubnubCoreBase.cs
@@ -823,7 +823,7 @@ namespace PubnubApi
                                                 Timetoken = pnFileResult.Timetoken,
                                                 Publisher = pnFileResult.Publisher,
                                             };
-                                            Dictionary<string, object> pnMsgObjDic = JsonDataParseInternalUtil.ConvertToDictionaryObject(pnFileResult.Message);
+                                            Dictionary<string, object> pnMsgObjDic = jsonLib.ConvertToDictionaryObject(pnFileResult.Message);
                                             if (pnMsgObjDic != null && pnMsgObjDic.Count > 0)
                                             {
                                                 if (pnMsgObjDic.ContainsKey("message") && pnMsgObjDic["message"]!= null)
@@ -832,7 +832,7 @@ namespace PubnubApi
                                                 }
                                                 if (pnMsgObjDic.ContainsKey("file"))
                                                 {
-                                                    Dictionary<string, object> fileObjDic = JsonDataParseInternalUtil.ConvertToDictionaryObject(pnMsgObjDic["file"]);
+                                                    Dictionary<string, object> fileObjDic = jsonLib.ConvertToDictionaryObject(pnMsgObjDic["file"]);
                                                     if (fileObjDic != null && fileObjDic.ContainsKey("id") && fileObjDic.ContainsKey("name"))
                                                     {
                                                         fileMessage.File = new PNFile { Id = fileObjDic["id"].ToString(), Name = fileObjDic["name"].ToString() };
