+++ src/Api/PubnubApiUnity/PubnubApiUnity.csproj
@@ -218,6 +218,9 @@
     </Compile>
 	<Compile Include="..\PubnubApi\Interface\ISubscribeOperation.cs" Link="Interface\ISubscribeOperation.cs" />
     <Compile Include="..\PubnubApi\Interface\IUnsubscribeOperation.cs" Link="Interface\IUnsubscribeOperation.cs" />
+    <Compile Include="..\PubnubApi\JsonDataParse\DeserializeToInternalObjectUtility.cs">
+      <Link>JsonDataParse\DeserializeToInternalObjectUtility.cs</Link>
+    </Compile>
 	<Compile Include="..\PubnubApi\RetryConfiguration.cs">
       <Link>RetryConfiguration.cs</Link>
     </Compile>
@@ -302,7 +305,6 @@
     <Compile Include="..\PubnubApi\Interface\IUrlRequestBuilder.cs">
       <Link>Interface\IUrlRequestBuilder.cs</Link>
     </Compile>
-    <Compile Include="..\PubnubApi\JsonDataParse\JsonDataParseInternalUtil.cs" Link="JsonDataParse\JsonDataParseInternalUtil.cs" />
     <Compile Include="..\PubnubApi\JsonDataParse\PNAccessManagerGrantJsonDataParse.cs" Link="JsonDataParse\PNAccessManagerGrantJsonDataParse.cs" />
     <Compile Include="..\PubnubApi\JsonDataParse\PNAddMessageActionJsonDataParse.cs" Link="JsonDataParse\PNAddMessageActionJsonDataParse.cs" />
     <Compile Include="..\PubnubApi\JsonDataParse\PNFetchHistoryJsonDataParse.cs" Link="JsonDataParse\PNFetchHistoryJsonDataParse.cs" />
@@ -980,7 +982,6 @@
     <Folder Include="Model\Derived\Presence\" />
     <Folder Include="Model\Server\" />
     <Folder Include="Helper\" />
-    <Folder Include="JsonDataParse\" />
     <Folder Include="Properties\" />
     <Folder Include="Push\Mpns\" />
     <Folder Include="Proxy\" />
